desafio 24/7

+++ src/controller/ProductController.js
@@ -152,7 +152,8 @@ export class ProductController {
       }
 
       try {
-        await productService.addProduct({ ...req.body });
+        const owner = req.session.user.email;
+        await productService.addProduct({ ...req.body, owner });
         let productList = await productService.getAllProducts();
         io.emit("updateProducts", productList);
         return res.json({ payload: `Product added` });
@@ -290,27 +291,40 @@ export class ProductController {
           ERROR_TYPES.INVALID_ARGUMENTS
         );
       }
-      try {
-        let products = await productService.deleteProduct(pid);
-        if (products.deletedCount > 0) {
-          let productList = await productService.getProductsPaginate();
-          io.emit("deleteProducts", productList);
-          return res.json({ payload: `Product ${pid} deleted` });
-        } else {
+      const userRole = req.session.user.role.toLowerCase();
+      const userEmail = req.session.user.email;
+      let product = await productService.getProductsBy({ _id: pid });
+
+      if (
+        userRole === "admin" ||
+        (userRole === "premium" && product.owner === userEmail)
+      ) {
+        try {
+          let products = await productService.deleteProduct(pid);
+          if (products.deletedCount > 0) {
+            let productList = await productService.getProductsPaginate();
+            io.emit("deleteProducts", productList);
+            return res.json({ payload: `Product ${pid} deleted` });
+          } else {
+            return CustomError.createError(
+              "ERROR",
+              null,
+              "Product not found",
+              ERROR_TYPES.NOT_FOUND
+            );
+          }
+        } catch (error) {
           return CustomError.createError(
             "ERROR",
             null,
-            "Product not found",
-            ERROR_TYPES.NOT_FOUND
+            `Error deleting product`,
+            ERROR_TYPES.INVALID_ARGUMENTS
           );
         }
-      } catch (error) {
-        return CustomError.createError(
-          "ERROR",
-          null,
-          `Error deleting product`,
-          ERROR_TYPES.INVALID_ARGUMENTS
-        );
+      } else {
+        return res
+          .status(403)
+          .json({ error: `Insufficient privileges to delete` });
       }
     } catch (error) {
       if (error.code !== 500) {
